+++ tests/Alley/WP/Alleyvate/Features/LoginNonceTest.php
@@ -19,10 +19,14 @@ namespace Alley\WP\Alleyvate\Features;
 use Mantle\Testing\Concerns\Refresh_Database;
 use Mantle\Testing\Exceptions\WP_Die_Exception;
 use Mantle\Testkit\Test_Case;
+use PHPUnit\Framework\Attributes\PreserveGlobalState;
+use PHPUnit\Framework\Attributes\RunClassInSeparateProcess;
 
 /**
  * Tests for the login nonce.
  */
+#[PreserveGlobalState( false )]
+#[RunClassInSeparateProcess]
 final class LoginNonceTest extends Test_Case {
 	use Refresh_Database;
 
@@ -32,19 +36,6 @@ final class LoginNonceTest extends Test_Case {
 	 * @var Login_Nonce
 	 */
 	private Login_Nonce $feature;
-
-	/**
-	 * Setup the test case.
-	 *
-	 * @param array ...$args The array of arguments passed to the class.
-	 */
-	public function __construct( ...$args ) {
-		parent::__construct( ...$args );
-
-		// Run the test in isolation to allow us to use http_response_code().
-		$this->setPreserveGlobalState( false );
-		$this->setRunClassInSeparateProcess( true );
-	}
 
 	/**
 	 * Set up.
@@ -82,13 +73,14 @@ final class LoginNonceTest extends Test_Case {
 
 		$pagenow = 'wp-login.php';
 
+		// Prevent the redirect and exit from firing.
+		add_filter( 'wp_redirect', fn ( $location ) => wp_die( esc_url( $location ) ) );
+
 		try {
 			Login_Nonce::action__pre_validate_login_nonce();
 		} catch ( WP_Die_Exception $e ) {
-			$this->assertSame( 'Login attempt failed. Please try again.', $e->getMessage() );
+			$this->assertSame( wp_login_url() . '?action=expired', $e->getMessage() );
 		}
-
-		$this->assertSame( 403, http_response_code() );
 	}
 
 	/**
@@ -165,21 +157,5 @@ final class LoginNonceTest extends Test_Case {
 
 		self::assertArrayHasKey( 'Cache-Control', $headers );
 		self::assertStringContainsString( 'no-store', $headers['Cache-Control'] );
-	}
-
-	/**
-	 * Verify that the no-store flag isn't added to other pages.
-	 */
-	public function test_non_login_page_is_stored() {
-		global $pagenow;
-
-		$pagenow = 'single.php'; // Anything other than wp-login.php.
-
-		$this->feature->boot();
-
-		$headers = wp_get_nocache_headers();
-
-		self::assertArrayHasKey( 'Cache-Control', $headers );
-		self::assertStringNotContainsString( 'no-store', $headers['Cache-Control'] );
 	}
 }